Introduce `suspenders:views` generator
Configures flash messages, page titles via the [title][] gem, and sets
the document [lang][].

[title]: https://github.com/calebhearth/title
[lang]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/lang

It should be noted that running `rails gscaffold` generates [views][]
that contain the flash. Although we could have [overridden][] this
generator, that would risk drift between our overridden generator and
the one in Rails core.

Additionally, we decided to remove the `FlashesHelper` introduced in
[6c562b95e2e3c7ea31a04e6ef71c6a44b0b9ab2f][] since that is not a pattern
we currently use.

[views]: https://github.com/rails/rails/blob/main/railties/lib/rails/generators/erb/scaffold/templates/index.html.erb.tt
[overridden]: https://guides.rubyonrails.org/generators.html#overriding-rails-generator-templates
[6c562b95e2e3c7ea31a04e6ef71c6a44b0b9ab2f]: https://github.com/thoughtbot/suspenders/commit/6c562b95e2e3c7ea31a04e6ef71c6a44b0b9ab2f