Remove option to wipe database
Running `dev:prime` calls `db:setup`, which appears to be wiping the
database for us. I tried updating it to use `db:prepare`, but that means
we will always seed the database with data from the `dev.rake` file,
resulting in duplicate records.

This implementation is consistent with the existing implementation.