Introduce `suspenders:tasks` generator
Relates to #1157

For now, this generator simply creates `dev.rake` which contains
[dev:prime][].

Note that we'll want this generator to run **after**
`suspenders:factories`.

If future application specific tasks need to be added, we can use this
generator. However, this should not be confused with the existing
pattern of creating tasks under the suspenders namespace, such as the
existing `lib/tasks/suspenders.rake`. Tasks under this namespace cannot
be edited by the consumer, whereas tasks generated by `suspenders:tasks`
are intended to be edited by the consumer.

[dev:prime]: https://thoughtbot.com/blog/priming-the-pump