Introduce `suspenders:setup` generator (#1157)
Replaces the [default setup script][] provided by Rails. The trade-off
being that the implementation is simple, but the cost is we risk
drifting from Rails. After attempting to use [gsub_file][] and
[insert_into_file][], it felt simpler to just override the file
completely.

Conditionally call [dev:prime][] to seed development data on top of [seed][]
data necessary for production. A follow-up commit will re-introduce this task.

Additionally, we [setup the test environment][] to avoid issues with
asset compilation.

[default setup script]: https://github.com/rails/rails/blob/main/railties/lib/rails/generators/rails/app/templates/bin/setup.tt
[gsub_file]: https://rubydoc.info/gems/thor/Thor/Actions#gsub_file-instance_method
[insert_into_file]: https://rubydoc.info/gems/thor/Thor/Actions#insert_into_file-instance_method
[dev:prime]: https://thoughtbot.com/blog/priming-the-pump
[seed]: https://guides.rubyonrails.org/active_record_migrations.html#migrations-and-seed-data
[setup the test environment]: https://github.com/rails/rails/pull/47719/files