Introduce `suspenders:db:migrate` task (#1161)
Lifted from our [dotfiles][], this task runs the latest migration, rolls
it back, and runs it again in an effort to ensure it's reversible. More
context can be found [in these comments][].

Unfortunately, the following implementation did not work because Rake recognized
that it `db:migrate` was already invoked, so it wasn't invoked a second time.

```ruby
task migrate: ["db:migrate", "db:rollback", "db:migrate", "db:test:prepare"]
```

Instead, we needed to `reenable` the task manually.

[dotfiles]: https://github.com/thoughtbot/dotfiles/blob/f149484269ef98e2bc80b7daa1988e214ddf8e8b/aliases#L12
[in these comments]: https://github.com/thoughtbot/dotfiles/commit/4882c418fd1cb4d052193d4e70a8df316e6cf9c6#r1933964