Styles Generator: Drop support for Tailwind
Follow-up to #1145 and #1148

We decided it's best to limit the decisions a consumer needs to make.
Even though we defaulted to PostCSS, providing an option to override
this value felt like it went against the Suspenders ethos.

Additionally, the current version of Suspenders uses PostCSS, so this
change aligns with current behavior.

Also removes linting rules around Tailwind.