Introduce `suspenders:cleanup:organize_gemfile` task
Introduce `Suspenders::Cleanup::OrganizeGemfile` class and corresponding
task in an effort to reduce duplicate groups in the modified Gemfile.

This is because the [gem_group][] method does not modify existing
groups. I've opened [#49512][] in an effort to fix this, but until then,
this task will suffice.

This class is designed to be run after `suspenders:install:web`, and
does not account for all edge cases. For example, it assumes gems are
grouped by symbols (i.e. :test and not "test"), and does not account for
inline syntax:

```ruby
gem 'my-gem', group: [:cucumber, :test]
```

We could consider extracting this into a Gem (with a fun name, of
course. Maybe "Polish"), but for now, I think this simple procedural
code if just fine.

Additionally, this commit removes duplicate Rake task that was generated
with the plugin.

[gem_group]: https://api.rubyonrails.org/classes/Rails/Generators/Actions.html#method-i-gem_group
[#49512]: https://github.com/rails/rails/pull/49512