Introduce `suspenders:test:environment` generator (#1182)
Configures test environment. This differs from #1156 in that this commit
is concerned with configuration, where that commit was concerned with
generating a holistic test suite. It's also possible to run each
generator independently, and the two should not rely on one another.

Disables [action_dispatch.show_exceptions][] in an effort to [improve
failure output][comment].

Enables [raise_on_missing_translations][] to keep parity with the same
setting in development #1149

[action_dispatch.show_exceptions]: https://edgeguides.rubyonrails.org/configuring.html#config-action-dispatch-show-exceptions
[comment]: https://github.com/thoughtbot/suspenders/pull/1149#discussion_r1424458925
[raise_on_missing_translations]: https://guides.rubyonrails.org/configuring.html#config-i18n-raise-on-missing-translations