Dynamically generate README
Create README based off the `desc` of the generators.

We chose to create a custom class instead of a traditional generator
because this script is intended to be run in the [application
template][] as a Rake task. It is not intended to be run as a standalone
generator, since it only makes sense to be run on concert with
`suspenders:install:web`.

Updates all generator descriptions to use Herdoc syntax to ensure
consistent line breaks.

A follow-up commit will update the descriptions to work better in a
README.

[application template]: https://guides.rubyonrails.org/rails_application_templates.html