Introduce `suspenders:install:web` generator and application template (#1152)
Create generator to invoke all necessary generators. We add it to the
`install` namespace to provide flexibility should we add other
installation options, such as ones for API Only applications.

Introduces [template][] as a means to invoke `suspenders:install:web`
when creating a new application. This serves as an alternative to the
system executable that was removed.

We call `db:prepare` in the template to ensure `bin/setup` works as
expected. This is because we overrode that file to use `dev:prime`,
which assume the database has been created.

[template]: https://guides.rubyonrails.org/rails_application_templates.html