Use `String#start_with?` in `OrganizeGemfile`
Closes #1204

It seems `active_support/core_ext/string` is not loaded when calling
`suspenders:cleanup:organize_gemfile` on applications **not** created
with Suspenders's [application template][].

We could require this file in the class, but instead, we simply use
`String#start_with?` to keep things simple.

Additionally, we ran `standard` to auto-lint the file.

[application template]: https://raw.githubusercontent.com/thoughtbot/suspenders/main/lib/install/web.rb