Prepare for "Tailored" release
With the merge of #1135 and some subsequent follow-ups, we are ready to
officially release the next version of Suspenders.

Because we moved to [calver][] in #1106, we need to continue using
calver, since the latest release `20230113.0` is greater than `3.0.0`,
which was set in haste in ab3eb9735eabd3c3827c7b6a9cce7d59e079559e

In an effort to better brand this release, we give it the code name
"Tailored".

[calver]: https://calver.org