Call dynamic-readme reusable workflow
We want to have a way to edit our README footer at one place
and have the changes from there be propagated to our repos.

By adding this snippet in the README, we call this reusable workflow:
https://github.com/thoughtbot/templates/blob/main/.github/workflows/dynamic-readme.yaml
that renders and updates the README footer dynamically.

More details about this approach can be found here:
https://thoughtbot.com/blog/automate-processes-with-reusable-github-workflows