README: Update instructions for generating a new app
Closes #1221

The [release of Rails 7.2.0][rails_7_2] introduced its [own linting
mechanism][rails-lint].

Unfortunately, this creates a conflict with [our linting
mechanism][suspenders-lint].

This is only relevant when running `rails new` with our [application
template][].

In order to fix this, we simply require the caller pass
`--skip_rubocop`.

A future commit could explore re-introducing the system executable to
account for our required options, but for now, this fixes the issue.

[rails_7_2]: https://guides.rubyonrails.org/v7.2/7_2_release_notes.html
[rails-lint]: https://guides.rubyonrails.org/v7.2/7_2_release_notes.html#add-omakase-rubocop-rules-by-default
[suspenders-lint]: https://github.com/thoughtbot/suspenders/blob/main/FEATURES.md#linting
[application template]: https://github.com/thoughtbot/suspenders/blob/main/lib/install/web.rb