Add support for SQLite
With increasing interest in SQLite across the Rails community as a
whole, it's valuable for us to start exploring it too. This expands the
existing checks for PostgreSQL to also include SQLite so that where we
want to try out SQLite we're able to. It also adjusts the CI workflow to
exclude PostgreSQL when generating a SQLite-based application.

https://rubyonrails.org/2024/11/7/rails-8-no-paas-required
https://fractaledmind.github.io/tags/sqlite/

Closes #1245