Introduce `development:seed` and `development:seed:replant` tasks
Since `db:seeds.rb` is intended to ensure

> ... the existence of records required to run the application in every
environment (production, development, test)

we have relied on [`dev:prime`][1] for loading data necessary for users
to view most of the features of the app in development.

However, there were a few areas of improvement.

First, rename `dev` namespace to `development`, and rename [`prime`][1]
task to `seed` for improved clarity.

The, introduces `development:seed:replant` take to create parity with
the existing `db:seed:replant` task.

Introduces `Seeder` class to encapsulate seed data, which is called in
the `development:seed` and `development:seed:replant` tasks. The
advantage to this is that the class can be called independently from the
task, and can also be [unit tested][2].

[1]: https://thoughtbot.com/blog/priming-the-pump
[2]: https://thoughtbot.com/blog/seeds-of-destruction