Load application.css instead of `:app`
`:app` is new in Rails 8 and has a special meaning for Propshaft. It is
not the same as `:application`.

In Propshaft, as shown here[1], `:app` loads all of the CSS files in
`app/assets`.. That causes an error with a the Suspenders setup, because
it loads `application.css` then tries (and fails) to load all of the
`@import`ed files as well.

The fix is to only load `"application.css"`.

[1]: https://github.com/rails/propshaft/blob/689e756689baa141291c7422f6a2df41b5ceead8/lib/propshaft/helper.rb#L13