Configure test suite
Lifts from the existing [testing][], [factories][] and [ci][]
generators, while also relying on the fact that some of these files will
already exist. For example, `.github/workflows/ci.yml` will be generated
by Rails, we just need to modify the file.

The previous version of Suspenders could be invoked on an **existing**
application, but this new version is intended to be used when creating
**new** applications. Since we're calling `rails new` with `--skip-test`
we know that the `.github/workflows/ci.yml` will be missing the test
steps.

Also, this commit enables all the "recommended" settings in
`spec_helper`, which is a change from the current version.

This commit also updates the library's CI script to account for the fact
that the generated applicaiton uses RSpec.

[testing]: https://github.com/thoughtbot/suspenders/blob/main/lib/generators/suspenders/ci_generator.rb
[factories]: https://github.com/thoughtbot/suspenders/blob/main/lib/generators/suspenders/ci_generator.rb
[ci]: https://github.com/thoughtbot/suspenders/blob/main/lib/generators/suspenders/ci_generator.rb