Improve CI script and fix broken layout
Prior to this commit, we had no specs to test. This was "fine", but it's
definitely better to have some real application code to test.

This commit updates the CI script by adding a basic static page, and
adding a corresponding system spec.

This revealed we had a broken layout, because we were including a
partial that does not exist.